[Unified Gating] Refactor gate code into gate.go file
